delete old spammers
