+++ crates/core/src/spammer/blockwise2.rs
@@ -59,11 +59,49 @@ where
                 .into_stream()
                 .flat_map(futures::stream::iter)
                 .map(|b| {
-                    println!("[[bw2]] block: {:?}", b);
+                    println!("new block: {:?}", b);
                     SpamTrigger::BlockHash(b)
                 })
                 .boxed();
             m
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::{
+        agent_controller::AgentStore,
+        db::MockDb,
+        generator::util::test::spawn_anvil,
+        spammer::util::test::{get_test_signers, MockCallback},
+        test_scenario::tests::MockConfig,
+    };
+
+    use super::*;
+
+    #[tokio::test]
+    async fn watches_blocks_and_spams_them() {
+        let anvil = spawn_anvil();
+        println!("anvil url: {}", anvil.endpoint_url());
+        let seed = crate::generator::RandSeed::from_str("444444444444");
+        let mut scenario = TestScenario::new(
+            MockConfig,
+            MockDb.into(),
+            anvil.endpoint_url(),
+            None,
+            seed,
+            get_test_signers().as_slice(),
+            AgentStore::new(),
+        )
+        .await
+        .unwrap();
+        let callback_handler = MockCallback;
+        let msg_handle = TxActorHandle::new(12, scenario.db.clone(), scenario.rpc_client.clone());
+        let spammer = BlockwiseSpammer2::new::<MockDb>(msg_handle, callback_handler);
+
+        let result = spammer.spam_rpc(&mut scenario, 10, 3, None).await;
+        println!("{:?}", result);
+        assert!(result.is_ok());
     }
 }
